+++ pkg/controller/operatingsystemconfig/generator/templates/script.suse-chost.template
