+++ pkg/controller/operatingsystemconfig/generator/testfiles/script/cloud-init
