+++ pkg/controller/operatingsystemconfig/generator/testfiles/script/script.memoryone-chost-reconcile
